requirements.txt
uvicorn
pymysql
fastapi
python-dotenv
absl-py==2.1.0
adagio==0.2.6
alembic==1.14.0
annotated-types==0.7.0
appdirs==1.4.4
arch==7.0.0
astunparse==1.6.3
autots==0.6.17
beartype==0.19.0
blinker==1.9.0
click==8.1.7
cloudpickle==3.1.0
cmdstanpy==1.2.4
colorlog==6.9.0
contourpy==1.3.0
coreforecast==0.0.15
cycler==0.12.1
Cython==3.0.11
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dtaidistance==2.3.12
dtw-python==1.5.3
duckdb==1.1.3
einops==0.8.0
esig==0.9.7
ETSformer-pytorch==0.1.1
filelock==3.13.1
filterpy==1.4.5
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.55.2
fs==2.4.16
fsspec==2024.2.0
fugue==0.9.1
gast==0.6.0
gateloop-transformer==0.2.5
gluonts==0.16.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.68.1
h5py==3.12.1
hmmlearn==0.3.3
holidays==0.62
huggingface-hub==0.27.0
iTransformer==0.6.0
itsdangerous==2.2.0
joblib==1.4.2
keras==3.7.0
keras-self-attention==0.51.0
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.43.0
Mako==1.3.8
Markdown==3.7
markdown-it-py==3.0.0
matplotlib==3.9.3
mdurl==0.1.2
ml-dtypes==0.3.2
mne==1.8.0
mpmath==1.3.0
namex==0.0.8
networkx==3.2.1
nfoursid==1.0.1
numba==0.60.0
numpy==1.26.4
opt_einsum==3.4.0
optree==0.13.1
optuna==4.1.0
pandas==2.2.1
patsy==1.0.1
pillow==11.0.0
plotly==5.24.1
pmdarima==2.0.4
polars==1.17.1
pooch==1.8.2
prophet==1.1.6
protobuf==4.25.5
pyaml==24.12.1
pyarrow==18.1.0
pycatch22==0.4.5
pydantic==2.10.3
pydantic_core==2.27.1
pyod==2.0.2
pyparsing==3.2.0
pyts==0.13.0
pywin32==307
retrying==1.3.4
rich==13.9.4
rotary-embedding-torch==0.8.6
scikit-base==0.11.0
scikit-learn==1.5.2
scikit-optimize==0.10.2
scikit-posthocs==0.11.1
scipy==1.13.1
seaborn==0.13.2
seasonal==0.3.1
shap==0.46.0
skforecast==0.13.0
skpro==2.7.0
sktime==0.34.1
slicer==0.0.8
SQLAlchemy==2.0.36
stanio==0.5.1
statsforecast==1.7.8
statsmodels==0.14.4
stumpy==1.13.0
sympy==1.13.1
tbats==1.1.3
tenacity==9.0.0
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
threadpoolctl==3.5.0
toolz==0.12.1
tqdm==4.67.1
triad==0.9.8
tsfresh==0.20.3
tslearn==0.6.3
tzdata==2024.2
u8darts==0.31.0
utilsforecast==0.2.10
Werkzeug==3.0.6
wrapt==1.17.0
xarray==2024.7.0
xgboost==2.1.3
zstandard==0.23.0